requirements.txt
langchain-core>=1.1.1
langgraph>=1.0.4
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
pandas>=2.3.3
python-dotenv>=1.0.0
